Full text search for AssortmentDictionary Name and Number column

+++ src/Application/Common/Models/AssortmentDictionary/AssortmentDictionaryDto.cs
@@ -15,7 +15,7 @@ public class AssortmentDictionaryDto
     public string? Kind { get; init; }
     public string? DN1 { get; init; }
     public string? DN2 { get; init; }
-    public double? Length { get; init; }
+    public ushort? Length { get; init; }
     public string? RO { get; init; }
     public string? NS { get; init; }
     public string? Comment { get; init; }
